--- src/utils/usedebounce.js
@@ -1,21 +1,0 @@
-import { useState, useEffect } from 'react';
-
-// Custom debounce hook
-function useDebounce(value, delay) {
-    const [debouncedValue, setDebouncedValue] = useState(value);
-
-    useEffect(() => {
-        const handler = setTimeout(() => {
-            setDebouncedValue(value);
-        }, delay);
-
-        // Clean up the timeout if value changes or component unmounts
-        return () => {
-            clearTimeout(handler);
-        };
-    }, [value, delay]);
-
-    return debouncedValue;
-}
-
-export default useDebounce;
